Add locks to C♯ completer
Add locks to prevent multiple starts of the OmniSharp server for the
same solution.

+++ ycmd/completers/cs/cs_completer.py
@@ -29,6 +29,7 @@ import requests
 import urlparse
 import logging
 import solutiondetection
+import threading
 
 SERVER_NOT_FOUND_MSG = ( 'OmniSharp server binary not found at {0}. ' +
                          'Did you compile it? You can do so by running ' +
@@ -89,6 +90,7 @@ class CsharpCompleter( Completer ):
     self._diagnostic_store = None
     self._max_diagnostics_to_display = user_options[
       'max_diagnostics_to_display' ]
+    self._solution_state_lock = threading.Lock()
 
     if not os.path.isfile( PATH_TO_OMNISHARP_BINARY ):
       raise RuntimeError(
@@ -108,12 +110,19 @@ class CsharpCompleter( Completer ):
 
 
   def _GetSolutionCompleter( self, request_data ):
+    """ Get the solution completer or create a new one if it does not already
+    exist. Use a lock to avoid creating the same solution completer multiple
+    times."""
     solution = self._GetSolutionFile( request_data[ "filepath" ] )
-    if not solution in self._completer_per_solution:
-      keep_logfiles = self.user_options[ 'server_keep_logfiles' ]
-      desired_omnisharp_port = self.user_options.get( 'csharp_server_port' )
-      completer = CsharpSolutionCompleter( solution, keep_logfiles, desired_omnisharp_port )
-      self._completer_per_solution[ solution ] = completer
+
+    with self._solution_state_lock:
+      if solution not in self._completer_per_solution:
+        keep_logfiles = self.user_options[ 'server_keep_logfiles' ]
+        desired_omnisharp_port = self.user_options.get( 'csharp_server_port' )
+        completer = CsharpSolutionCompleter( solution,
+                                             keep_logfiles,
+                                             desired_omnisharp_port )
+        self._completer_per_solution[ solution ] = completer
 
     return self._completer_per_solution[ solution ]
 
@@ -337,7 +346,8 @@ class CsharpSolutionCompleter:
     self._filename_stdout = None
     self._omnisharp_port = None
     self._omnisharp_phandle = None
-    self._desired_omnisharp_port = desired_omnisharp_port;
+    self._desired_omnisharp_port = desired_omnisharp_port
+    self._server_state_lock = threading.RLock()
 
 
   def CodeCheck( self, request_data ):
@@ -350,60 +360,67 @@ class CsharpSolutionCompleter:
 
 
   def _StartServer( self ):
-    """ Start the OmniSharp server """
-    self._logger.info( 'startup' )
+    """ Start the OmniSharp server if not already running. Use a lock to avoid
+    starting the server multiple times for the same solution. """
+    with self._server_state_lock:
+      if self.ServerIsRunning():
+        return
 
-    path_to_solutionfile = self._solution_path
-    self._logger.info(
-        u'Loading solution file {0}'.format( path_to_solutionfile ) )
+      self._logger.info( 'Starting OmniSharp server' )
 
-    self._ChooseOmnisharpPort()
+      path_to_solutionfile = self._solution_path
+      self._logger.info(
+          u'Loading solution file {0}'.format( path_to_solutionfile ) )
 
-    command = [ PATH_TO_OMNISHARP_BINARY,
-                '-p',
-                str( self._omnisharp_port ),
-                '-s',
-                u'{0}'.format( path_to_solutionfile ) ]
+      self._ChooseOmnisharpPort()
 
-    if not utils.OnWindows() and not utils.OnCygwin():
-      command.insert( 0, 'mono' )
+      command = [ PATH_TO_OMNISHARP_BINARY,
+                  '-p',
+                  str( self._omnisharp_port ),
+                  '-s',
+                  u'{0}'.format( path_to_solutionfile ) ]
 
-    if utils.OnCygwin():
-      command.extend( [ '--client-path-mode', 'Cygwin' ] )
+      if not utils.OnWindows() and not utils.OnCygwin():
+        command.insert( 0, 'mono' )
 
-    filename_format = os.path.join( utils.PathToTempDir(),
-                                    u'omnisharp_{port}_{sln}_{std}.log' )
+      if utils.OnCygwin():
+        command.extend( [ '--client-path-mode', 'Cygwin' ] )
 
-    solutionfile = os.path.basename( path_to_solutionfile )
-    self._filename_stdout = filename_format.format(
-        port = self._omnisharp_port, sln = solutionfile, std = 'stdout' )
-    self._filename_stderr = filename_format.format(
-        port = self._omnisharp_port, sln = solutionfile, std = 'stderr' )
+      filename_format = os.path.join( utils.PathToTempDir(),
+                                      u'omnisharp_{port}_{sln}_{std}.log' )
 
-    with open( self._filename_stderr, 'w' ) as fstderr:
-      with open( self._filename_stdout, 'w' ) as fstdout:
-        self._omnisharp_phandle = utils.SafePopen(
-            command, stdout = fstdout, stderr = fstderr )
+      solutionfile = os.path.basename( path_to_solutionfile )
+      self._filename_stdout = filename_format.format(
+          port = self._omnisharp_port, sln = solutionfile, std = 'stdout' )
+      self._filename_stderr = filename_format.format(
+          port = self._omnisharp_port, sln = solutionfile, std = 'stderr' )
 
-    self._solution_path = path_to_solutionfile
+      with open( self._filename_stderr, 'w' ) as fstderr:
+        with open( self._filename_stdout, 'w' ) as fstdout:
+          self._omnisharp_phandle = utils.SafePopen(
+              command, stdout = fstdout, stderr = fstderr )
 
-    self._logger.info( 'Starting OmniSharp server' )
+      self._solution_path = path_to_solutionfile
 
 
   def _StopServer( self ):
-    """ Stop the OmniSharp server """
-    self._logger.info( 'Stopping OmniSharp server' )
+    """ Stop the OmniSharp server using a lock. """
+    with self._server_state_lock:
+      if not self.ServerIsRunning():
+        return
 
-    self._TryToStopServer()
+      self._logger.info( 'Stopping OmniSharp server' )
 
-    # Kill it if it's still up
-    if self.ServerIsRunning():
-      self._logger.info( 'Killing OmniSharp server' )
-      self._omnisharp_phandle.kill()
+      self._TryToStopServer()
 
-    self._CleanupAfterServerStop()
+      # Kill it if it's still up
+      if self.ServerIsRunning():
+        self._logger.info( 'Killing OmniSharp server' )
+        self._omnisharp_phandle.kill()
 
-    self._logger.info( 'Stopped OmniSharp server' )
+      self._CleanupAfterServerStop()
+
+      self._logger.info( 'Stopped OmniSharp server' )
 
 
   def _TryToStopServer( self ):
@@ -423,16 +440,16 @@ class CsharpSolutionCompleter:
     self._omnisharp_phandle = None
     if ( not self._keep_logfiles ):
       if self._filename_stdout:
-        os.unlink( self._filename_stdout );
+        os.unlink( self._filename_stdout )
       if self._filename_stderr:
-        os.unlink( self._filename_stderr );
+        os.unlink( self._filename_stderr )
 
 
-  def _RestartServer ( self ):
-    """ Restarts the OmniSharp server """
-    if self.ServerIsRunning():
+  def _RestartServer( self ):
+    """ Restarts the OmniSharp server using a lock. """
+    with self._server_state_lock:
       self._StopServer()
-    return self._StartServer()
+      self._StartServer()
 
 
   def _ReloadSolution( self ):
